requirements.txt
feedparser==6.0.10
schedule==1.2.0
pandas==2.1.0
fpdf==1.7.2
slack_sdk==3.23.0
gpt4all==1.1.6
python-dotenv==1.0.1